pyproject.toml
[project]
name = "plsl"
version = "0.1.0"
description = ""
authors = [
    {name = "Patrice Fortin",email = "patricefortin.x@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "pyqtgraph (>=0.13.7,<0.14.0)",
    "pyside6 (>=6.8.2.1,<7.0.0.0)",
    "pylsl (>=1.17.6,<2.0.0)",
    "mne (>=1.9.0,<2.0.0)",
    "h5py (>=3.13.0,<4.0.0)",
    "pyserial (>=3.5,<4.0)",
    "pyxdf (>=1.17.0,<2.0.0)",
    "python-rtmidi (>=1.5.8,<2.0.0)",
    "mido (>=1.3.3,<2.0.0)",
]

[tool.poetry]
packages = [{include = "plsl", from = "src"}]


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
